requirements.txt
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
ConfigArgParse==0.14.0
coverage==4.5.4
Flask==1.1.1
flask-restplus==0.13.0
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
locustio==0.11.0
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.1
packaging==19.1
pluggy==0.13.0
py==1.8.0
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
pytest-cov==2.7.1
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
